safari-600.7-branch/LayoutTests/touchadjustment/nested-shadow-node-expected.txt

Test the case where a clickable target contains a shadow-DOM element. The adjusted point should snap to the location of the shadow-DOM element if close enough to the original touch position.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS adjustedNode.id is element.id
PASS adjustedPoint.x is within 10 of 88
PASS adjustedPoint.y is within 10 of 28
PASS adjustedNode.id is element.id
PASS adjustedPoint.x is within 10 of 88
PASS adjustedPoint.y is within 10 of 28
PASS adjustedNode.id is element.id
PASS adjustedPoint.x is within 10 of 88
PASS adjustedPoint.y is within 10 of 28
PASS adjustedNode.id is element.id
PASS adjustedPoint.x is within 10 of 88
PASS adjustedPoint.y is within 10 of 28
PASS adjustedNode.id is element.id
PASS adjustedPoint.x is within 1 of 58
PASS adjustedPoint.y is within 1 of 58
PASS successfullyParsed is true

TEST COMPLETE
